modules/data_source/datasource_to_hdfs/requirements.txt
pyDataCanvas >= 0.4.6,<0.5.0
boto
sarge
